+++ program/src/modele/algobrut.cpp
@@ -15,7 +15,7 @@ void AlgoBrut::generateSols(int index,vector<Scale*> solutionPossible)
         solutionsPossibles.push_back(solutionPossible);
     else
     {
-        for (int i=0;i<KpartiteGraph[index].size();i++)
+        for (size_t i=0;i<KpartiteGraph[index].size();i++)
         {
             solutionPossible.push_back(KpartiteGraph[index][i]);
             generateSols(index+1,solutionPossible);
@@ -32,20 +32,20 @@ vector<vector<Scale*> > AlgoBrut::findLeastsConsecutivesNotesChanges()
     vector<int> values;
     int value;
 
-    for (int i=0;i<solutionsPossibles.size();i++)//on remplit le tableau values
+    for (size_t i=0;i<solutionsPossibles.size();i++)//on remplit le tableau values
     {
         value=0;
-        for (int j=0;j<solutionsPossibles[i].size()-1;j++)//on fait la somme des differences de notes entre chaque couple de gammes consécutives
+        for (size_t j=0;j<solutionsPossibles[i].size()-1;j++)//on fait la somme des differences de notes entre chaque couple de gammes consécutives
             value+=solutionsPossibles[i][j]->notesDifferencesWithScale(solutionsPossibles[i][j+1]);
         values.push_back(value);
     }
 
     int minValue=INFINITY;
-    for (int i=0;i<values.size();i++)//on cherche la plus petite valeur possible
+    for (size_t i=0;i<values.size();i++)//on cherche la plus petite valeur possible
         minValue=min(values[i],minValue);
 
     vector<vector<Scale*> >res;
-    for (int i=0;i<values.size();i++)//on ajoute dans le résultat retour toutes les occurences de la contrainte optimisée
+    for (size_t i=0;i<values.size();i++)//on ajoute dans le résultat retour toutes les occurences de la contrainte optimisée
         if (values[i]==minValue)
             res.push_back(solutionsPossibles[i]);
 
@@ -58,21 +58,21 @@ vector<vector<Scale*> > AlgoBrut::findLeastsConsecutivesScalesChanges()
     vector<int> values;
     int value;
 
-    for (int i=0;i<solutionsPossibles.size();i++)
+    for (size_t i=0;i<solutionsPossibles.size();i++)
     {
         value=0;
-        for (int j=0;j<solutionsPossibles[i].size()-1;j++)
+        for (size_t j=0;j<solutionsPossibles[i].size()-1;j++)
             if (!solutionsPossibles[i][j]->equals(solutionsPossibles[i][j+1]))
                 value++;
         values.push_back(value);
     }
 
     int minValue=INFINITY;
-    for (int i=0;i<values.size();i++)
+    for (size_t i=0;i<values.size();i++)
         minValue=min(minValue,values[i]);
 
     vector<vector<Scale*> > res;
-    for (int i=0;i<values.size();i++)
+    for (size_t i=0;i<values.size();i++)
         if (values[i]==minValue)
             res.push_back(solutionsPossibles[i]);
 
@@ -84,10 +84,10 @@ vector<vector<Scale*> > AlgoBrut::findLeastsTotalScales()
     vector<int> values;
     vector<Scale*> distinctsScales;//sert à stocker les différentes gammes de chaque suite de gammes
 
-    for (int i=0;i<solutionsPossibles.size();i++)
+    for (size_t i=0;i<solutionsPossibles.size();i++)
     {
 
-        for (int j=0;j<solutionsPossibles[i].size();j++)
+        for (size_t j=0;j<solutionsPossibles[i].size();j++)
             if (!isScaleInScales(solutionsPossibles[i][j],distinctsScales))
                 distinctsScales.push_back(solutionsPossibles[i][j]);
 
@@ -96,11 +96,11 @@ vector<vector<Scale*> > AlgoBrut::findLeastsTotalScales()
     }
 
     int minValue=INFINITY;
-    for (int i=0;i<values.size();i++)
+    for (size_t i=0;i<values.size();i++)
         minValue=min(minValue,values[i]);
 
     vector<vector<Scale*> > res;
-    for (int i=0;i<values.size();i++)
+    for (size_t i=0;i<values.size();i++)
         if (values[i]==minValue)
             res.push_back(solutionsPossibles[i]);
